Updated to last biicode features and moved to standard GL folder
