+++ .gitignore
@@ -7,3 +7,7 @@ Thumbs.db
 
 #Mac DS_Store file
 .DS_Stor
+
+#Biicode base layout
+bii
+bin
