+++ biicode.conf
@@ -1,0 +1,46 @@
+# Biicode configuration file
+
+[requirements]
+
+
+[parent]
+    # The parent version of this block. Must match folder name. E.g.
+    # user/block  # No version number means not published yet
+    # You can change it to publish to a different track, and change version, e.g.
+    # user/block(track): 7
+
+[paths]
+    # Local directories to look for headers (within block)
+    # /
+    include
+
+[dependencies]
+    # Manual adjust file implicit dependencies, add (+), remove (-), or overwrite (=)
+    # hello.h + hello_imp.cpp hello_imp2.cpp
+    # *.h + *.cpp
+    include/GL/glew.h + src/glew.c
+
+[mains]
+    # Manual adjust of files that define an executable
+    # !main.cpp  # Do not build executable from this file
+    # main2.cpp # Build it (it doesnt have a main() function, but maybe it includes it)
+
+[tests]
+    # Manual adjust of files that define a CTest test
+    # test/* pattern to evaluate this test/ folder sources like tests
+
+[hooks]
+    # These are defined equal to [dependencies],files names matching bii*stage*hook.py
+    # will be launched as python scripts at stage = {post_process, clean}
+    # CMakeLists.txt + bii/my_post_process1_hook.py bii_clean_hook.py
+
+[includes]
+    # Mapping of include patterns to external blocks
+    # hello*.h: user3/depblock  # includes will be processed as user3/depblock/hello*.h
+
+[data]
+    # Manually define data files dependencies, that will be copied to bin for execution
+    # By default they are copied to bin/user/block/... which should be taken into account
+    # when loading from disk such data
+    # image.cpp + image.jpg  # code should write open("user/block/image.jpg")
+
