+++ src/glew.c
@@ -31,12 +31,12 @@
 ** THE POSSIBILITY OF SUCH DAMAGE.
 */
 
-#include "glew.h"
+#include "GL/glew.h"
 
 #if defined(_WIN32)
-#  include "wglew.h"
+#  include "GL/wglew.h"
 #elif !defined(__ANDROID__) && !defined(__native_client__) && !defined(__HAIKU__) && (!defined(__APPLE__) || defined(GLEW_APPLE_GLX))
-#  include "glxew.h"
+#  include "GL/glxew.h"
 #endif
 
 #include <stddef.h>  /* For size_t */
